refactor(layout): compose parent-relative geometry once
Separate private source positions from actual IR content origins and make block displacements signed. Reparent only moved roots; remove cumulative source margins from owned bodies. Migrate CLI and TUI together, including cell-local content, nonparagraph markers, hard term roots, zero-width targets and closed schema decoding.

Share marker collision and final-padding rules. Cross-review added negative-container, link-coordinate, trailing-anchor and extreme-origin regressions, and corrected exact-half source rounding plus bounded rebase arithmetic. Update the unpublished v0.11 schema and consumer documentation.

Validation: 199 explanation gold queries and 51 semantic fixtures clean; all-feature workspace targets exercised, three obsolete presentation assertions updated and rerun; focused geometry, protocol and UI suites plus strict workspace Clippy pass. Full release gate will be rerun after the remaining source-layout slices.

+++ crates/mant-engine/examples/roff_structure_profile.rs
@@ -65,7 +65,7 @@ struct IrStructure {
     definition_items: usize,
     table_rows: usize,
     table_spanning_cells: usize,
-    max_indent_columns: u16,
+    max_indent_columns: i32,
     hard_breaks: usize,
     manual_links: usize,
     external_links: usize,
